# Optimization on `mergedStringIter`