We are calling `Err()` recursively on every call to the `Next()` method. This is not really needed and can cause a huge overhead when merging multiples blocks into one.

https://github.com/prometheus/prometheus/blob/6c008ec56af958debf4123d7cd58a6681d17303d/tsdb/compact.go#L736

Instead of calling `Err()` for every `Next()` call we can just after calling the next on the inners iterators.

Let's have the following scenario:

```

         C
        / \
       F   \
            G
           / \
          H   I
```

Before, calling C.Next() would call `Err()` recursively for all child nodes of C (F, G, H and I) even though the it would only call the `Next()` for nodes on one path (EX: C ->  G -> H).
With this change, we are not recusing over all child of C anymore, instead we are calling `Err()` only when we call the `Next()` on the child node and storing the result. This result is now returned when `c.Err()` is called.

Benchmark Results:

```
name                 old time/op    new time/op    delta
MergedStringIter-12     4.13s ± 0%     0.36s ± 0%  -91.38%  (p=0.016 n=5+4)

name                 old alloc/op   new alloc/op   delta
MergedStringIter-12    5.62MB ± 0%    2.95MB ± 0%  -47.52%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
MergedStringIter-12      300k ± 0%      167k ± 0%  -44.40%  (p=0.008 n=5+5)
```

Cortex Compactor  cpu flame-graph while creating the symbols table:

![Screen Shot 2023-03-13 at 4 58 28 PM](https://user-images.githubusercontent.com/4027760/224857833-2462ef05-f9c8-4d9a-a333-7b0b8e0922b5.png)