@bboreham 2023-03-14
Good catch on the perf issue, but surely the bigger problem is we are implementing K-Way Merge as K-1 instances of 2-way merge?

There are already multiple implementations in Prometheus using heap as a priority queue, plus I have a new implementation which is very cool.