@bboreham 2023-03-14
I guess Prometheus compactor is expecting to merge ~4 blocks into 1, whereas Cortex and others are merging dozens, so the problem shows up more.