@alanprot 2023-03-14
Yeah.. but what h suggest? Implementing k-way merge here or can we ship this for now and implement the k-way merge in a next pr? I’m ok either way..